Disable JS post-processing on existing blocks
The "resume" template includes JS code to render "blocks" (entries) on the client side.
This makes it simpler to write such entries, by simply writing a level-3 section in Markdown, with the 1st paragraph being a sub-title, the 2nd being the location, the 3rd being the dates, and subsequent paragraphs being details.
The correct HTML code is produced by parsing this structure and re-arranging elements, on the client side.

However, the same code could be produced directly from R code: this allows more flexibility (i.e., adding chips, choosing the CSS classes, etc.).
This was not initially possible, as the JS code re-arranges everything that is a level-3 section.
This commit simply makes the JS code skip sections that already have the `blocks` CSS class, indicating that already is in the desired structure.
Markdown-based sections do not have this class, and will thus still be re-formatted.