Refactor the contact functions to use htmltools
We previously printed Glue templates, which works-ish, but is not robust, and not flexible.
In particular, we could not easily add linebreaks (`<br />`) between multiple calls to
`cv.contact`, Rmarkdown/Pandoc would escape them if we tried to print them as-is.

Instead, we now return a htmltool tag. A single call will be printed as expected; multiple
calls can be enclosed in a htmltool tagList. Linebreaks can be added to this tagList if necessary.