Add the icons package for more available icons
Fontawesome is already available through the `<i class=...>` syntax, but some icons
were missing, especially for the Skills section (programming languages, tools, ...).
Simpleicons has these icons, but is not as easy to use as Fontawesome; the `icons`
package simplifies their usage.

The `icons` package is only available through GitHub, so we must use `remotes` to
install it.

Note that it requires to download the icons first; the deploy workflow is updated to
do so, but any dev environment must also download them to render locally.